Refactor loading of paywall preview resources into its own file.
